TelegramCallBot-main/requirements.txt
python-dotenv==1.0.1
apscheduler==3.10.4
aiogram==3.13.1
fastapi==0.115.0
uvicorn[standard]==0.32.0
setuptools==70.3.0
wheel==0.44.0

